refactor: waiting 패키지 분리
- Waiting과 Reservation은 서로 다르고, 어느 한 쪽이 다른 쪽의 하위 도메인도 아니라고 판단했다.
따라서 Waiting 관련 클래스를 `waiting`으로 분리했다.
