+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.member.service.dto.LoginMemberInfo;
 import roomescape.reservation.controller.dto.AdminReservationCreateRequest;
-import roomescape.reservation.controller.dto.MemberReservationResponse;
+import roomescape.reservation.controller.dto.MyReservationResponse;
 import roomescape.reservation.controller.dto.ReservationCreateRequest;
 import roomescape.reservation.controller.dto.ReservationResponse;
 import roomescape.reservation.controller.dto.ReservationSearchConditionRequest;
@@ -64,10 +64,10 @@ public class ReservationController {
     }
 
     @GetMapping("/reservations-mine")
-    public ResponseEntity<List<MemberReservationResponse>> findAllMine(LoginMemberInfo loginMemberInfo) {
+    public ResponseEntity<List<MyReservationResponse>> findAllMine(LoginMemberInfo loginMemberInfo) {
         final List<ReservationInfo> reservationInfos = reservationService.findReservationsByMemberId(loginMemberInfo.id());
-        final List<MemberReservationResponse> responses = reservationInfos.stream()
-                .map(MemberReservationResponse::new)
+        final List<MyReservationResponse> responses = reservationInfos.stream()
+                .map(MyReservationResponse::new)
                 .toList();
         return ResponseEntity.ok().body(responses);
     }
