+++ src/main/java/roomescape/reservation/controller/dto/MyReservationResponse.java
@@ -4,8 +4,9 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import roomescape.reservation.service.dto.ReservationInfo;
+import roomescape.waiting.service.dto.WaitingInfo;
 
-public record MemberReservationResponse(
+public record MyReservationResponse(
         long reservationId,
         String theme,
         LocalDate date,
@@ -14,12 +15,21 @@ public record MemberReservationResponse(
         String status
 ) {
 
-    public MemberReservationResponse(final ReservationInfo reservationInfo) {
+    public MyReservationResponse(final ReservationInfo reservationInfo) {
         this(reservationInfo.id(),
                 reservationInfo.theme().name(),
                 reservationInfo.date(),
                 reservationInfo.time().startAt(),
-                reservationInfo.status().getDisplayName()
+                "예약"
+        );
+    }
+
+    public MyReservationResponse(final WaitingInfo waitingInfo) {
+        this(waitingInfo.id(),
+                waitingInfo.theme().name(),
+                waitingInfo.date(),
+                waitingInfo.time().startAt(),
+                "%d번째 예약대기".formatted(waitingInfo.order())
         );
     }
 }
